+++ src/components/Header.js
@@ -1,13 +1,21 @@
 // https://avatars.githubusercontent.com/HarshitJoshi9152
 import React, { useEffect, useRef, useState } from "react";
-import { useHistory, useParams, Link, useRouteMatch} from "react-router-dom";
+import { useHistory, useParams, Link, useRouteMatch } from "react-router-dom";
 import { useHotkeys } from "react-hotkeys-hook";
-import Loader from "react-loader";
-import {Img} from 'react-image';
+// import Loader from "react-loader";
+import { Img } from "react-image";
 
-import { atom, useAtom } from 'jotai'
+// import { atom, useAtom } from "jotai";
 
-import {VscColorMode, VscInfo, VscGithubInverted, VscAccount, VscSettingsGear} from "react-icons/vsc";
+import RightNav from "./rightNav.js";
+
+import {
+	VscColorMode,
+	VscInfo,
+	VscGithubInverted,
+	VscAccount,
+	VscSettingsGear
+} from "react-icons/vsc";
 
 import { makeFriendly } from "./../utils/num.js";
 import "./header.css";
@@ -21,46 +29,46 @@ const repoLink = "https://github.com/HarshitJoshi9152/showerthoughts";
 --fg: #C5C8C6;
 todo: add seen var switch
 */
-const r = document.querySelector(':root');
+const r = document.querySelector(":root");
 const bgLight = "#edf6f9";
 const fgLight = "#1c1412";
 const bgDark = "#1d1f21";
 const fgDark = "#C5C8C6";
 
 const iconStyle = {
-		width: "32px",
-		height: "32px",
-		margin: "0px 8px 0px 8px"
+	width: "32px",
+	height: "32px",
+	margin: "0px 8px 0px 8px"
 };
 
 const toggleTheme = () => {
-		const rs = getComputedStyle(r);
-		const primary = rs.getPropertyValue('--primary');
-		const text = rs.getPropertyValue('--text');
+	const rs = getComputedStyle(r);
+	const primary = rs.getPropertyValue("--primary");
+	const text = rs.getPropertyValue("--text");
 
-		// alert(primary);
-		// alert(text);
+	// alert(primary);
+	// alert(text);
 
-		// todo: why does this not work on the first try ?
-		if (primary === bgLight && text === fgLight) {
-			// even tho it seems to be true (not actually) this doesnt seem to be triggered
-			r.style.setProperty('--primary', bgDark);
-			r.style.setProperty('--text', fgDark);
-		} else {
-			// for 1st try because it is in light mode by default.
-			// todo: wont work after customisations tho lol
-			r.style.setProperty('--primary', bgLight);
-			r.style.setProperty('--text', fgLight);
-		}
-		// else if (primary === bgDark && text === fgDark){
-		// 	r.style.setProperty('--primary', bgLight);
-		// 	r.style.setProperty('--text', fgLight);
-		// } 
-}
+	// todo: why does this not work on the first try ?
+	if (primary === bgLight && text === fgLight) {
+		// even tho it seems to be true (not actually) this doesnt seem to be triggered
+		r.style.setProperty("--primary", bgDark);
+		r.style.setProperty("--text", fgDark);
+	} else {
+		// for 1st try because it is in light mode by default.
+		// todo: wont work after customisations tho lol
+		r.style.setProperty("--primary", bgLight);
+		r.style.setProperty("--text", fgLight);
+	}
+	// else if (primary === bgDark && text === fgDark){
+	// 	r.style.setProperty('--primary', bgLight);
+	// 	r.style.setProperty('--text', fgLight);
+	// }
+};
 
 // for 1st try it doesnt work, but it does after that so.
 // todo: wont work after customisations tho lol
-toggleTheme()
+toggleTheme();
 
 export default function Header() {
 	const { subreddit } = useParams();
@@ -72,18 +80,14 @@ export default function Header() {
 
 	const sel_subreddit = (sub) => {
 		closeSelectMenu();
-		
+
 		if (subreddit === sub.slice(2)) {
-			// maybe hisroy.goBack(); idk1
-			// wtf is this all that had to be done ???? wtf wow
-			// and it doesnt even call another posts fetch and fixed the bad overlapping wow
-			alert("smae sub ewont change a thing lol");
-			// history.replace(`/${subreddit}`);
+			// history.replace(`/${subreddit}`); // note: wont cause useEffect to run if param is in dependencies.
 			return null;
 		}
 
 		// sub includes "r/"
-		history.push("/" + sub.slice(2), {prevSub: subreddit});
+		history.push("/" + sub.slice(2), { prevSub: subreddit });
 		// todo: get rid of the slice use the input value margin left we saw on stackOverflow the other day.
 	};
 
@@ -93,6 +97,10 @@ export default function Header() {
 	const [desc, setDesc] = useState();
 	const img = useRef();
 	const [loaded, setLoaded] = useState(false);
+
+	// check localStorage or JWT for default value.
+	const [loggedIn, setLoggedIn] = useState(false);
+
 	useHotkeys("/", toggleSelectMenu);
 	// why cant i use this here ?
 	// useHotKeys("ctrl + /", () => alert("show all keyboard shortcuts"))
@@ -106,9 +114,8 @@ export default function Header() {
 				if (!res.ok) {
 					// {data} is undefined in this case.
 					throw Error("Not Found");
-				}
-				else {
-					return res.json()
+				} else {
+					return res.json();
 				}
 			})
 			.then(({ data }) => {
@@ -117,16 +124,16 @@ export default function Header() {
 				setImgSrc(data?.icon_img || data?.community_icon);
 				setSubCount(data?.subscribers);
 				setActiveCount(data?.active_user_count);
-				setDesc(data?.public_description)
+				setDesc(data?.public_description);
 				setLoaded(true);
 				// todo: dont show the icon if we cant load it.
 			})
 			.catch((e) => {
 				console.log("Header", e.message);
-				if(e.message === "Not Found") {
-					// 
-				};
-			})
+				if (e.message === "Not Found") {
+					//
+				}
+			});
 		return () => {
 			// reset on sub change
 			setLoaded(false);
@@ -135,13 +142,13 @@ export default function Header() {
 			setActiveCount(null);
 			// do i really need to do this ? not really tho
 			// setImgSrc("");
-		}
+		};
 	}, [subreddit]);
 	// todo: public_description_html vs public_description
 
 	const toggleInfo = () => {
-		alert("niuce")
-	}
+		alert("niuce");
+	};
 
 	useHotkeys("t", toggleTheme);
 
@@ -149,10 +156,12 @@ export default function Header() {
 		<header>
 			{/*welcome to The Open Source reddit client focused on browsing{" "}*/}
 			<span>
-				<span onClick={() => {
-					if(!match.isExact)
-						history.push("/" + subreddit)
-				}} title={desc}>
+				<span
+					onClick={() => {
+						if (!match.isExact) history.push("/" + subreddit);
+					}}
+					title={desc}
+				>
 					{loaded && <Img className="sub-icon" src={imgSrc} />}
 					<p className="banner">r/{subreddit}</p>
 				</span>
@@ -179,19 +188,7 @@ export default function Header() {
 					></SubredditSelect>
 				)}
 			</span>
-			<span className="right">
-				<VscAccount style={iconStyle} title="Login"/>
-				<VscSettingsGear style={iconStyle} title="Login"/>
-				<span className="theme-switch">
-					<VscColorMode style={iconStyle} onClick={toggleTheme}/>
-				</span>
-				<span className="info">
-					<VscInfo style={iconStyle} onClick={toggleInfo}/>
-				</span>
-				<a href={repoLink} target="_blank" rel="noreferrer" className="repo-link">
-					<VscGithubInverted style={iconStyle} title={repoLink}/>
-				</a>
-			</span>
+			<RightNav {...{ toggleTheme, toggleInfo, loggedIn, setLoggedIn }} />
 		</header>
 	);
 }