+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 ## CHANGELOG
 
-1. changed default value for subreddit to r/all.
-2. fixed the StackView overlapping issue.
-3. changed name to Zenitsu.
+1. Refactored Header component.
+2. added login popup UI.